fix: move shield button to address bar
