+++ src/renderer/views/app/components/SiteButtons/index.tsx
@@ -7,8 +7,9 @@ import {
   ICON_KEY,
   ICON_MAGNIFY_PLUS,
   ICON_MAGNIFY_MINUS,
+  ICON_SHIELD,
 } from '~/renderer/constants/icons';
-import { ipcRenderer } from 'electron';
+import { ipcRenderer, remote } from 'electron';
 import { parse } from 'url';
 import store from '../../store';
 import { ToolbarButton } from '../ToolbarButton';
@@ -62,13 +63,31 @@ ipcRenderer.on('zoom-factor-updated', (e, zoomFactor, showDialog) => {
   }
 });
 
+const onShieldContextMenu = (e: React.MouseEvent<HTMLDivElement>) => {
+  const menu = remote.Menu.buildFromTemplate([
+    {
+      checked: store.settings.object.shield,
+      label: 'Enabled',
+      type: 'checkbox',
+      click: () => {
+        store.settings.object.shield = !store.settings.object.shield;
+        store.settings.save();
+      },
+    },
+  ]);
+
+  menu.popup();
+};
+
 export const SiteButtons = observer(() => {
   const { selectedTab } = store.tabs;
 
   let hasCredentials = false;
+  let blockedAds = 0;
 
   if (selectedTab) {
     hasCredentials = selectedTab.hasCredentials;
+    blockedAds = selectedTab.blockedAds;
   }
 
   const dense = !store.isCompact;
@@ -101,6 +120,14 @@ export const SiteButtons = observer(() => {
         dense={dense}
         onMouseDown={onStarClick}
       />
+      <ToolbarButton
+        size={16}
+        badge={store.settings.object.shield && blockedAds > 0}
+        badgeText={blockedAds.toString()}
+        icon={ICON_SHIELD}
+        opacity={store.settings.object.shield ? 0.87 : 0.54}
+        onContextMenu={onShieldContextMenu}
+      ></ToolbarButton>
     </>
   );
 });
